Update blogpost.md
I stopped at line 58. 

I think I will need to meet with you to talk about what the intention of this blog posts is. Who it is targeted to and what the expectations are for edits. 

It may be better to have in a google doc so I can drop some comments in sections where I have a question.