don't store or print samples
fixes #6
fixes #7

samples can be very large, for example the known bug of (bex (bex 30))
in jam.
to accommodate testing even very large (but easily expressible samples)
we only store a (reasonably collission resistant) hash of the samples to
check for duplicates
and we return the defying sample (if found) instead of printing it.
printing is optional: the previous behavior is unchanged.
instead it is not possible to call another entry-point which returns the
result, `check-verbose`.
the default implementation `check` still prints just like before.